refactor - 앱 모듈 패키지명 변경
