+++ app/src/main/java/com/hugg/presentation/module/NetworkModule.kt
@@ -1,4 +1,4 @@
-package com.hugg.app.module
+package com.hugg.presentation.module
 
 import android.util.Log
 import com.google.gson.GsonBuilder
