+++ app/src/main/java/com/hugg/presentation/module/RepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.hugg.app.module
+package com.hugg.presentation.module
 
 import com.hugg.data.repository.*
 import com.hugg.domain.repository.*
