+++ feature/account/src/main/java/com/hugg/account/accountMain/AccountScreen.kt
@@ -40,6 +40,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.PathEffect
@@ -107,7 +108,7 @@ import com.hugg.feature.util.TimeFormatter
 import com.hugg.feature.util.UnitFormatter
 import kotlinx.coroutines.flow.distinctUntilChanged
 
-@OptIn(ExperimentalMaterial3Api::class)
+@OptIn(ExperimentalComposeUiApi::class, ExperimentalMaterial3Api::class)
 @Composable
 fun AccountContainer(
     navigateToSubsidyList : (Int) -> Unit = {},
